+++ src/xyz/oihana/schema/constants/traits/ThesaurusTrait.php
@@ -4,6 +4,7 @@ namespace xyz\oihana\schema\constants\traits;
 
 use xyz\oihana\schema\constants\traits\thesaurus\ConceptSchemeTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ConceptTrait;
+use xyz\oihana\schema\constants\traits\thesaurus\SkosMappingsTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\SkosNotesTrait;
 use xyz\oihana\schema\constants\traits\thesaurus\ThesaurusTermTrait;
 
@@ -18,6 +19,7 @@ trait ThesaurusTrait
 {
     use ConceptSchemeTrait ,
         ConceptTrait ,
+        SkosMappingsTrait ,
         SkosNotesTrait ,
         ThesaurusTermTrait ;
 }